v0.7.1: Fix float32 dtype issues and add comprehensive tests
- Enforce float64 dtype throughout package to prevent float32 issues
  - backend.py: Set default dtype=np.float64 for zeros()
  - FIF.py: Replace np.full_like() with B.full_like() to use backend
  - fbm.py: Replace np.real() with B.real() for consistency
  - spectral.py: Add explicit dtype=np.float64 to np.logspace() and np.array()
  - haar_fluctuation.py: Add explicit dtype=np.float64 to np.ones() and np.array()
- Add comprehensive unit tests (tests/test_basic_functionality.py)
  - Tests all simulation methods (fBm_1D, fBm_ND, FIF_1D, FIF_ND)
  - Tests all analysis methods (structure function, Haar, spectral)
  - Validates dtype consistency (ensures float64 output)
  - 75 tests covering wide parameter ranges
- Bump version to 0.7.1 in pyproject.toml, setup.py, and __init__.py

+++ scaleinvariance/analysis/haar_fluctuation.py
@@ -74,9 +74,9 @@ def haar_fluctuation_analysis(data, order=1, max_sep=None, axis=0, lags='powers 
         if lag % 2 != 0 or lag > max_sep:
             haar_flucs.append(np.nan)
             continue
-            
+
         # Construct Haar kernel
-        kernel = np.ones(lag) / (lag/2)
+        kernel = np.ones(lag, dtype=np.float64) / (lag/2)
         kernel[:lag//2] *= -1
 
         # Embed the 1D kernel into an n-D kernel with its values along the specified axis.
@@ -97,7 +97,7 @@ def haar_fluctuation_analysis(data, order=1, max_sep=None, axis=0, lags='powers 
         mean_absolute_haar_fluctuation = np.nanmean(abs_conv)
         haar_flucs.append(mean_absolute_haar_fluctuation)
 
-    return np.array(lags), np.array(haar_flucs)
+    return np.array(lags, dtype=np.float64), np.array(haar_flucs, dtype=np.float64)
 
 
 def haar_fluctuation_hurst(data, min_sep=None, max_sep=None, axis=0, return_fit=False):
